refactor settings files into /config
