+++ app/src/androidTest/java/com/github/fribourgsdp/radio/MainActivityTest.kt
@@ -9,6 +9,7 @@ import androidx.test.espresso.intent.matcher.IntentMatchers
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.github.fribourgsdp.radio.config.SettingsActivity
 import com.github.fribourgsdp.radio.data.view.UserProfileActivity
 import junit.framework.Assert.assertEquals
 import org.hamcrest.Matchers
