+++ app/src/androidTest/java/com/github/fribourgsdp/radio/SettingsBuilderTest.kt
@@ -1,5 +1,7 @@
 package com.github.fribourgsdp.radio
 
+import com.github.fribourgsdp.radio.config.Settings
+import com.github.fribourgsdp.radio.config.SettingsBuilder
 import org.junit.Assert
 import org.junit.Test
