+++ app/src/androidTest/java/com/github/fribourgsdp/radio/SettingsTest.kt
@@ -2,6 +2,7 @@ package com.github.fribourgsdp.radio
 
 import android.content.Context
 import androidx.test.core.app.ApplicationProvider
+import com.github.fribourgsdp.radio.config.Settings
 import org.junit.Assert
 import org.junit.Test
